Change compiler plugin to code generator
