--- misc/testerina/modules/testerina-compiler-plugin/src/main/java/org/ballerinalang/testerina/compiler/MainFlowGenerationTask.java
@@ -1,312 +1,0 @@
-/*
- *  Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.ballerinalang.testerina.compiler;
-
-import io.ballerina.compiler.syntax.tree.CaptureBindingPatternNode;
-import io.ballerina.compiler.syntax.tree.CheckExpressionNode;
-import io.ballerina.compiler.syntax.tree.ExpressionNode;
-import io.ballerina.compiler.syntax.tree.FunctionArgumentNode;
-import io.ballerina.compiler.syntax.tree.FunctionBodyBlockNode;
-import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
-import io.ballerina.compiler.syntax.tree.FunctionSignatureNode;
-import io.ballerina.compiler.syntax.tree.ImportDeclarationNode;
-import io.ballerina.compiler.syntax.tree.MapTypeDescriptorNode;
-import io.ballerina.compiler.syntax.tree.MethodCallExpressionNode;
-import io.ballerina.compiler.syntax.tree.ModuleMemberDeclarationNode;
-import io.ballerina.compiler.syntax.tree.ModulePartNode;
-import io.ballerina.compiler.syntax.tree.Node;
-import io.ballerina.compiler.syntax.tree.NodeFactory;
-import io.ballerina.compiler.syntax.tree.NodeList;
-import io.ballerina.compiler.syntax.tree.OptionalTypeDescriptorNode;
-import io.ballerina.compiler.syntax.tree.PositionalArgumentNode;
-import io.ballerina.compiler.syntax.tree.ReturnTypeDescriptorNode;
-import io.ballerina.compiler.syntax.tree.SeparatedNodeList;
-import io.ballerina.compiler.syntax.tree.SimpleNameReferenceNode;
-import io.ballerina.compiler.syntax.tree.StatementNode;
-import io.ballerina.compiler.syntax.tree.SyntaxKind;
-import io.ballerina.compiler.syntax.tree.Token;
-import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
-import io.ballerina.compiler.syntax.tree.TypedBindingPatternNode;
-import io.ballerina.projects.Document;
-import io.ballerina.projects.DocumentConfig;
-import io.ballerina.projects.DocumentId;
-import io.ballerina.projects.Module;
-import io.ballerina.projects.ModuleId;
-import io.ballerina.projects.Project;
-import io.ballerina.projects.plugins.AnalysisTask;
-import io.ballerina.projects.plugins.CompilationAnalysisContext;
-
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-/**
- * Code analyzer to generate the Testerina runtime function.
- */
-public class MainFlowGenerationTask implements AnalysisTask<CompilationAnalysisContext> {
-
-    private static final String TEST_EXEC_BAL = "test_execute.bal";
-    private static final String TEST_REGISTER_FUNCTION = "registerTest";
-    private static final String TEST_EXECUTE_FUNCTION = "execute";
-    private static final String TEST_REPORT_FUNCTION = "report";
-    private static final String TEST_ORG_NAME = "ballerina";
-    private static final String TEST_MODULE_NAME = "test";
-    private static final String TEST_EXEC_FUNCTION = "main";
-    private static final String TEST_DIRECTORY = "tests/";
-
-    @Override
-    public void perform(CompilationAnalysisContext compilationAnalysisContext) {
-
-        for (ModuleId moduleId : compilationAnalysisContext.currentPackage().moduleIds()) {
-            Module module = compilationAnalysisContext.currentPackage().module(moduleId);
-            DocumentConfig documentConfig = generateDocument(
-                    compilationAnalysisContext.currentPackage().project(), module);
-            module.modify().addTestDocument(documentConfig).apply();
-        }
-    }
-
-    private static DocumentConfig generateDocument(Project project, Module module) {
-
-        List<StatementNode> statements = new ArrayList<>();
-
-        // Register all the test cases of the module
-        for (DocumentId documentId : module.testDocumentIds()) {
-            Document document = module.document(documentId);
-            Node node = document.syntaxTree().rootNode();
-            TestFunctionVisitor testFunctionVisitor = new TestFunctionVisitor();
-            node.accept(testFunctionVisitor);
-
-            // Add the statements, 'check test:registerTest(<name>, <function>);'
-            testFunctionVisitor.getTestStaticFunctions().forEach(func -> {
-                statements.add(invokeRegisterFunction(func.functionName().toString(), func.functionName().toString()));
-            });
-
-            // Add the statements, 'check <function>()'
-            testFunctionVisitor.getTestDynamicFunctions().forEach(func -> {
-                statements.add(invokeFactoryFunction(func.functionName().toString(),
-                        func.functionSignature().returnTypeDesc()));
-            });
-        }
-
-        // Add the statement, 'map<error?> result = test:execute();'
-        MapTypeDescriptorNode resultDescriptor = NodeFactory.createMapTypeDescriptorNode(
-                NodeFactory.createToken(SyntaxKind.MAP_KEYWORD),
-                NodeFactory.createTypeParameterNode(
-                        NodeFactory.createToken(SyntaxKind.LT_TOKEN),
-                        NodeFactory.createOptionalTypeDescriptorNode(
-                                NodeFactory.createIdentifierToken(SyntaxKind.ERROR_KEYWORD.stringValue()),
-                                NodeFactory.createToken(SyntaxKind.QUESTION_MARK_TOKEN)
-                        ),
-                        NodeFactory.createToken(SyntaxKind.GT_TOKEN)
-                )
-        );
-        statements.add(getFunctionCallStatement("result", resultDescriptor,
-                getTestFunctionCall(TEST_EXECUTE_FUNCTION, NodeFactory.createSeparatedNodeList(new ArrayList<>()))));
-
-        // Add the statement, 'test:report(result);'
-        statements.add(getFunctionCallStatement(getTestFunctionCall(TEST_REPORT_FUNCTION,
-                NodeFactory.createSeparatedNodeList(List.of(NodeFactory.createIdentifierToken("result"))))));
-
-        // Construct the test execution function
-        FunctionBodyBlockNode functionBodyNode = NodeFactory.createFunctionBodyBlockNode(
-                NodeFactory.createToken(
-                        SyntaxKind.OPEN_BRACE_TOKEN,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae("\n"))),
-                null,
-                NodeFactory.createNodeList(statements.toArray(new StatementNode[0])),
-                NodeFactory.createToken(SyntaxKind.CLOSE_BRACE_TOKEN));
-        FunctionDefinitionNode functionDefinition = NodeFactory.createFunctionDefinitionNode(
-                SyntaxKind.FUNCTION_DEFINITION,
-                null,
-                NodeFactory.createNodeList(NodeFactory.createToken(
-                        SyntaxKind.PUBLIC_KEYWORD,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" ")))),
-                NodeFactory.createToken(
-                        SyntaxKind.FUNCTION_KEYWORD,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" "))),
-                NodeFactory.createIdentifierToken(TEST_EXEC_FUNCTION,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList()),
-                NodeFactory.createEmptyNodeList(),
-                getFunctionSignature(),
-                functionBodyNode);
-
-        // Construct the module part node
-        // Add the line, 'import ballerina/test;'
-        ImportDeclarationNode testImport = NodeFactory.createImportDeclarationNode(
-                NodeFactory.createToken(
-                        SyntaxKind.IMPORT_KEYWORD,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" "))),
-                NodeFactory.createImportOrgNameNode(
-                        NodeFactory.createIdentifierToken(TEST_ORG_NAME),
-                        NodeFactory.createToken(SyntaxKind.SLASH_TOKEN)),
-                NodeFactory.createSeparatedNodeList(
-                        NodeFactory.createIdentifierToken(TEST_MODULE_NAME)),
-                null,
-                NodeFactory.createToken(
-                        SyntaxKind.SEMICOLON_TOKEN,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae("\n"))));
-
-        // Construct the document content
-        NodeList<ModuleMemberDeclarationNode> nodeList = NodeFactory.createNodeList(List.of(functionDefinition));
-        Token eofToken = NodeFactory.createToken(SyntaxKind.EOF_TOKEN, NodeFactory.createEmptyMinutiaeList(),
-                NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae("\n")));
-        ModulePartNode modulePartNode = NodeFactory.createModulePartNode(
-                NodeFactory.createNodeList(testImport), nodeList, eofToken);
-
-        // Construct the document config
-        String newFileContent = modulePartNode.toSourceCode();
-        Path filePath = module.isDefaultModule()
-                ? project.sourceRoot().resolve(TEST_DIRECTORY).resolve(TEST_EXEC_BAL)
-                : project.sourceRoot().resolve("modules").resolve(module.moduleName().moduleNamePart())
-                .resolve(TEST_DIRECTORY).resolve(TEST_EXEC_BAL);
-        DocumentId newDocumentId = DocumentId.create(filePath.toString(), module.moduleId());
-        return DocumentConfig.from(newDocumentId, newFileContent, TEST_DIRECTORY + TEST_EXEC_BAL);
-    }
-
-    private static StatementNode invokeRegisterFunction(String testNameVal, String testFuncVal) {
-
-        PositionalArgumentNode testName = NodeFactory.createPositionalArgumentNode(
-                NodeFactory.createBasicLiteralNode(
-                        SyntaxKind.STRING_LITERAL,
-                        NodeFactory.createLiteralValueToken(
-                                SyntaxKind.STRING_LITERAL_TOKEN,
-                                "\"" + testNameVal + "\"",
-                                NodeFactory.createEmptyMinutiaeList(),
-                                NodeFactory.createEmptyMinutiaeList())));
-        PositionalArgumentNode testFunction = NodeFactory.createPositionalArgumentNode(
-                NodeFactory.createSimpleNameReferenceNode(NodeFactory.createIdentifierToken(testFuncVal)));
-
-        SeparatedNodeList<FunctionArgumentNode> separatedNodeList = getFunctionParamList(testName, testFunction);
-
-        return getFunctionCallStatement(getCheckedExpressionStatement(getTestFunctionCall(
-                TEST_REGISTER_FUNCTION, separatedNodeList)));
-    }
-
-    private static StatementNode invokeFactoryFunction(String functionName,
-                                                       Optional<ReturnTypeDescriptorNode> returnTypeDesc) {
-
-        SimpleNameReferenceNode simpleNameReferenceNode =
-                NodeFactory.createSimpleNameReferenceNode(NodeFactory.createIdentifierToken(functionName));
-        ExpressionNode functionExpressionNode = NodeFactory.createFunctionCallExpressionNode(
-                simpleNameReferenceNode,
-                NodeFactory.createToken(SyntaxKind.OPEN_PAREN_TOKEN),
-                NodeFactory.createSeparatedNodeList(new ArrayList<>()),
-                NodeFactory.createToken(SyntaxKind.CLOSE_PAREN_TOKEN));
-
-        return getFunctionCallStatement(returnTypeDesc.isPresent()
-                ? getCheckedExpressionStatement(functionExpressionNode)
-                : functionExpressionNode);
-    }
-
-    public static MethodCallExpressionNode getTestFunctionCall(String functionName,
-                                                               SeparatedNodeList<FunctionArgumentNode> nodeList) {
-
-        SimpleNameReferenceNode functionRefNode = NodeFactory.createSimpleNameReferenceNode(
-                NodeFactory.createIdentifierToken(functionName));
-        SimpleNameReferenceNode testRefNode = NodeFactory.createSimpleNameReferenceNode(
-                NodeFactory.createIdentifierToken(TEST_MODULE_NAME));
-        return NodeFactory.createMethodCallExpressionNode(
-                testRefNode,
-                NodeFactory.createToken(SyntaxKind.COLON_TOKEN),
-                functionRefNode,
-                NodeFactory.createToken(SyntaxKind.OPEN_PAREN_TOKEN),
-                nodeList,
-                NodeFactory.createToken(SyntaxKind.CLOSE_PAREN_TOKEN));
-    }
-
-    public static StatementNode getFunctionCallStatement(ExpressionNode expression) {
-
-        return NodeFactory.createExpressionStatementNode(SyntaxKind.FUNCTION_CALL, expression,
-                NodeFactory.createToken(SyntaxKind.SEMICOLON_TOKEN, NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae("\n"))));
-    }
-
-    public static StatementNode getFunctionCallStatement(String varName, TypeDescriptorNode typeDescriptorNode,
-                                                         ExpressionNode expression) {
-
-        CaptureBindingPatternNode captureBindingPatternNode = NodeFactory.createCaptureBindingPatternNode(
-                NodeFactory.createIdentifierToken(
-                        varName,
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" ")),
-                        NodeFactory.createEmptyMinutiaeList()));
-        TypedBindingPatternNode typedBindingPatternNode = NodeFactory.createTypedBindingPatternNode(
-                typeDescriptorNode, captureBindingPatternNode);
-        return NodeFactory.createVariableDeclarationNode(
-                NodeFactory.createEmptyNodeList(),
-                null,
-                typedBindingPatternNode,
-                NodeFactory.createToken(SyntaxKind.EQUAL_TOKEN),
-                expression,
-                NodeFactory.createToken(
-                        SyntaxKind.SEMICOLON_TOKEN,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae("\n"))));
-    }
-
-    private static CheckExpressionNode getCheckedExpressionStatement(ExpressionNode expression) {
-
-        return NodeFactory.createCheckExpressionNode(
-                SyntaxKind.CHECK_EXPRESSION,
-                NodeFactory.createToken(SyntaxKind.CHECK_KEYWORD,
-                        NodeFactory.createEmptyMinutiaeList(),
-                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" "))),
-                expression);
-    }
-
-    private static SeparatedNodeList<FunctionArgumentNode> getFunctionParamList(PositionalArgumentNode... args) {
-
-        List<Node> nodeList = new ArrayList<>();
-        for (PositionalArgumentNode arg : args) {
-            nodeList.add(arg);
-            nodeList.add(NodeFactory.createToken(SyntaxKind.COMMA_TOKEN));
-        }
-        if (args.length > 0) {
-            nodeList.remove(nodeList.size() - 1);
-        }
-        return NodeFactory.createSeparatedNodeList(nodeList);
-    }
-
-    private static FunctionSignatureNode getFunctionSignature() {
-
-        OptionalTypeDescriptorNode optionalErrorTypeDescriptorNode =
-                NodeFactory.createOptionalTypeDescriptorNode(
-                        NodeFactory.createParameterizedTypeDescriptorNode(
-                                SyntaxKind.ERROR_TYPE_DESC, NodeFactory.createToken(SyntaxKind.ERROR_KEYWORD), null),
-                        NodeFactory.createToken(SyntaxKind.QUESTION_MARK_TOKEN, NodeFactory.createEmptyMinutiaeList(),
-                                NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" "))));
-
-        ReturnTypeDescriptorNode returnTypeDescriptorNode =
-                NodeFactory.createReturnTypeDescriptorNode(NodeFactory
-                                .createToken(SyntaxKind.RETURNS_KEYWORD, NodeFactory.createMinutiaeList(
-                                                NodeFactory.createWhitespaceMinutiae(" ")),
-                                        NodeFactory.createMinutiaeList(NodeFactory.createWhitespaceMinutiae(" "))),
-                        NodeFactory.createEmptyNodeList(), optionalErrorTypeDescriptorNode);
-
-        return NodeFactory.createFunctionSignatureNode(NodeFactory.createToken(SyntaxKind.OPEN_PAREN_TOKEN),
-                NodeFactory.createSeparatedNodeList(),
-                NodeFactory.createToken(SyntaxKind.CLOSE_PAREN_TOKEN), returnTypeDescriptorNode);
-    }
-}
